requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
itsdangerous==2.1.2
python-dotenv==0.21.1
Flask-Migrate==3.1.0
Flask-Reuploaded==1.2.0
psycopg2-binary
pytest==7.0.1
python-dotenv==0.21.1
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Werkzeug==2.2.3

# this was causing errors?
# mysqlclient==2.1.1
